requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch==2.0.0
torchvision==0.15.1
fairscale
sentencepiece
Pillow
opencv-python
gradio
tqdm
git+https://github.com/csuhan/timm_0_3_2.git
git+https://github.com/openai/CLIP.git